+++ CP3Music/Helpers/Helpers.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 // modulo function which supports negative numbers
-func mod(_ a: Int, _ n: Int) -> Int {
+func mod(_ a: Int8, _ n: Int8) -> Int8 {
     precondition(n > 0, "modulus must be positive")
     let r = a % n
     return r >= 0 ? r : r + n
